Switch scanner to use NIO for performance; small API changes.
- Switch recursive scanning to use an NIO TreeWalker rather than custom
  recursive scanning code. This should be faster than java.util.zip when
  scanning jars.
- Small API changes (will be documented in the 1.10.0 release).

+++ src/main/java/io/github/lukehutch/fastclasspathscanner/matchprocessor/FileMatchProcessor.java
@@ -45,5 +45,5 @@ public interface FileMatchProcessor {
      * @param lengthBytes
      *            The length of the InputStream in bytes.
      */
-    public void processMatch(String relativePath, InputStream inputStream, int lengthBytes) throws IOException;
+    public void processMatch(String relativePath, InputStream inputStream, long lengthBytes) throws IOException;
 }
